feat: Support C# Decimal

+++ src/CSharpToJava.Core/Context/TypeMappingService.cs
@@ -765,7 +765,7 @@ public class TypeMappingService
             "ushort"  => "short",
             "float"   => "float",
             "double"  => "double",
-            "decimal" => "double",
+            "decimal" => "Decimal",
             "char"    => "char",
             "void"    => "void",
             "object"  => "Object",
